requirements.txt
django==1.8.4
django_smart_selects==1.1.1

